normal op in rethinkdb

+++ server/src/endpoint/remove.ts
@@ -1,13 +1,23 @@
 import * as r from 'rethinkdb';
+import * as fbs from '../msg_generated';
 
-export async function remove({ collection, selector }, collections, send, errorHandle, dbConnection) {
+export async function remove(base: fbs.Base, collections, send, errorHandle, dbConnection) {
   try {
+    const msg = new fbs.Remove();
+    base.msg(msg);
+    const collection = msg.collection();
+
+    let selector = msg.selector();
+    if (selector) {
+      selector = JSON.parse(selector);
+    }
+
     const conn = dbConnection.connection();
     const table = collections.get(collection).table;
     let result: r.WriteResult;
     if (typeof selector === 'string') {
       result = await (table.get(selector) as any).delete().run(conn);
-    } else if (typeof selector === 'object') {
+    } else if (selector !== null && typeof selector === 'object') {
       result = await table.filter(selector).delete().run(conn);
     } else {
       result = await table.delete().run(conn);
@@ -23,7 +33,7 @@ export async function remove({ collection, selector }, collections, send, errorH
     }
 
     send({
-      state: 'complete',
+      done: true,
       data
     })
   } catch (e) {
